regen with std::min and std::max-aware sql2cpp

+++ src/pluto_security/Table_Alert.cpp
@@ -422,7 +422,7 @@ if (is_null[3])
 return "NULL";
 
 char *buf = new char[39];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_DetectionTime.c_str(), (unsigned long) min(19,m_DetectionTime.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_DetectionTime.c_str(), (unsigned long) min((size_t)19,m_DetectionTime.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -436,7 +436,7 @@ if (is_null[4])
 return "NULL";
 
 char *buf = new char[39];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_ExpirationTime.c_str(), (unsigned long) min(19,m_ExpirationTime.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_ExpirationTime.c_str(), (unsigned long) min((size_t)19,m_ExpirationTime.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -476,7 +476,7 @@ if (is_null[7])
 return "NULL";
 
 char *buf = new char[39];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_ResetTime.c_str(), (unsigned long) min(19,m_ResetTime.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_ResetTime.c_str(), (unsigned long) min((size_t)19,m_ResetTime.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -594,7 +594,7 @@ if (is_null[16])
 return "NULL";
 
 char *buf = new char[29];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_psc_mod.c_str(), (unsigned long) min(14,m_psc_mod.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_psc_mod.c_str(), (unsigned long) min((size_t)14,m_psc_mod.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -673,7 +673,7 @@ values_list_comma_separated = values_list_comma_separated + pRow->PK_Alert_asSQL
 		{
 			
 			
-			long int id	= (long int) mysql_insert_id(database->m_pMySQL);
+			long int id = (long int) mysql_insert_id(database->m_pMySQL);
 		
 			if (id!=0)
 pRow->m_PK_Alert=id;
