+++ src/pluto_security/Table_Alert_Device.cpp
@@ -295,7 +295,7 @@ if (is_null[3])
 return "NULL";
 
 char *buf = new char[39];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_DetectionTime.c_str(), (unsigned long) min(19,m_DetectionTime.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_DetectionTime.c_str(), (unsigned long) min((size_t)19,m_DetectionTime.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -361,7 +361,7 @@ if (is_null[8])
 return "NULL";
 
 char *buf = new char[29];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_psc_mod.c_str(), (unsigned long) min(14,m_psc_mod.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_psc_mod.c_str(), (unsigned long) min((size_t)14,m_psc_mod.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -440,7 +440,7 @@ values_list_comma_separated = values_list_comma_separated + pRow->PK_Alert_Devic
 		{
 			
 			
-			long int id	= (long int) mysql_insert_id(database->m_pMySQL);
+			long int id = (long int) mysql_insert_id(database->m_pMySQL);
 		
 			if (id!=0)
 pRow->m_PK_Alert_Device=id;
