+++ src/pluto_security/Table_ModeChange.cpp
@@ -309,7 +309,7 @@ if (is_null[3])
 return "NULL";
 
 char *buf = new char[39];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_ChangeTime.c_str(), (unsigned long) min(19,m_ChangeTime.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_ChangeTime.c_str(), (unsigned long) min((size_t)19,m_ChangeTime.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -388,7 +388,7 @@ if (is_null[9])
 return "NULL";
 
 char *buf = new char[29];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_psc_mod.c_str(), (unsigned long) min(14,m_psc_mod.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_psc_mod.c_str(), (unsigned long) min((size_t)14,m_psc_mod.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -467,7 +467,7 @@ values_list_comma_separated = values_list_comma_separated + pRow->PK_ModeChange_
 		{
 			
 			
-			long int id	= (long int) mysql_insert_id(database->m_pMySQL);
+			long int id = (long int) mysql_insert_id(database->m_pMySQL);
 		
 			if (id!=0)
 pRow->m_PK_ModeChange=id;
