+++ src/pluto_security/Table_psc_security_schema.cpp
@@ -169,7 +169,7 @@ if (is_null[1])
 return "NULL";
 
 char *buf = new char[131071];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_Value.c_str(), (unsigned long) min(65535,m_Value.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_Value.c_str(), (unsigned long) min((size_t)65535,m_Value.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -235,7 +235,7 @@ values_list_comma_separated = values_list_comma_separated + pRow->PK_psc_securit
 		{
 			
 			
-			long int id	= (long int) mysql_insert_id(database->m_pMySQL);
+			long int id = (long int) mysql_insert_id(database->m_pMySQL);
 		
 			if (id!=0)
 pRow->m_PK_psc_security_schema=id;
